refactor generation command
