+++ modelina-cli/src/index.ts
@@ -1,1 +1,3 @@
 export {run} from '@oclif/core'
+
+export {generateModels, ModelinaArgs, ModelinaFlags, Languages} from './helpers/generate'
